requirements.txt
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
blacksheep==2.0.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.1.0.post1
essentials==1.1.5
essentials-openapi==1.0.9
fastapi==0.106.0
flasgger==0.9.7.1
Flask==3.0.0
guardpost==1.0.2
h11==0.14.0
httptools==0.6.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
MarkupSafe==2.1.3
marshmallow==3.20.1
mistune==3.0.2
mongoengine==0.27.0
packaging==23.2
pika==1.3.2
pydantic==2.5.2
pydantic-openapi-schema==1.5.1
pydantic-settings==2.1.0
pydantic_core==2.14.5
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rodi==2.0.5
rpds-py==0.15.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
Werkzeug==3.0.1
